Fix: クラス名をMemoに変え、別ファイルに切り分けた
